function_requirements.txt
functions-framework==3.*
google-generativeai==0.8.5